Feat/#187

+++ src/main/java/org/plteco/ploytechcourse/domain/user/mypage/service/MyPage.java
@@ -8,7 +8,7 @@ public interface MyPage {
     Long getNumberOfPosts(Long id);
     Long getNumberOfLove(Long id);
     String getProfile(Long id);
-    Long getGeneration(Long id);
+    Integer getGeneration(Long id);
     String getUid(Long id);
     RoleEnum getRole(Long id);
 }
